add task wrappers
Task wrappers are a functional alternative to hooks for changing the
behavior of tasks. You can add something like this to your project.clj:

    :task-wrappers {help (fn [f]
                           (fn [project & args]
                             (prn "foo")
                             (apply f project args)))
                    version leiningen.wrappers/wrap-version}

This wraps the help and version tasks, using an anonymous function for
the former and a namespaced function for the latter:

    (defn wrap-version [f]
      (fn [project]
        (prn "bar")
        (f project)))

NOTE: This is just a proof of concept at this point.